+++ Private/Find-Type.ps1
@@ -1,13 +1,47 @@
 function Find-Type {
+    [CmdletBinding(DefaultParameterSetName = 'Predicate')]
     param (
+        [Parameter(
+            Mandatory,
+            Position = 0,
+            ValueFromPipeline
+        )]
         [string]
         $TypeName,
 
+        [Parameter(
+            Position = 1,
+            ValueFromPipeline,
+            ParameterSetName = 'Predicate'
+        )]
+        [scriptblock]
+        $Predicate,
+
+        [Parameter(
+            Position = 1,
+            ParameterSetName = 'Exact'
+        )]
         [switch]
         $Exact
     )
 
     process {
+        <# GetTypes().Where({
+            $_.IsPublic
+            -and
+            (
+                $_.FullName -eq $TypeName
+                -or
+                $_.FullName -match "[\w.]+\.${TypeName}$"
+                -or
+                (
+                    $_.Name -eq $TypeName
+                    -and
+                    $_.Namespace -eq 'System'
+                )
+            )
+            })
+        #>
         try {
             $AllAssemblies = [appdomain]::CurrentDomain.GetAssemblies()
 
@@ -24,14 +58,17 @@ function Find-Type {
                     }
                 }
             }
+            elseif($Predicate) {
+                # Look for the type name using the provided predicate scriptblock
+                $Type = $AllAssemblies.GetTypes() | Where-Object -FilterScript $Predicate
+            }
             else {
                 # Look for the type name (fuzzy search) in loaded assemblies
-                $Type = $AllAssemblies.GetTypes() | Where-Object {
+                $Type = $AllAssemblies.GetTypes() | Where-Object -FilterScript {
                     $_.IsPublic -and
                     (
                         $_.FullName -like $TypeName -or
-                        $_.FullName -match "[\w.]+\.${TypeName}$" -or
-                        $_.Namespace -eq 'System'
+                        $_.FullName -match "[\w.]+\.${TypeName}$"
                     )
                 }
             }