+++ CHANGELOG.md
@@ -2,8 +2,34 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
-and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [0.5.5] - 2025-10-10
+
+### Added
+- `New-ScriptExtent` private function for creating custom extents for use with DiagnosticRecords
+
+### Changed
+- Adjusted `Measure-CheckParamBlockParen` extent
+
+## [0.4.5] - 2025-10-09
+
+### Added
+- `Measure-CheckKeywordSpacing` rule scaffolded, but not complete
+- Added further [TODO](TODO.md) entries for rule ideas
+- Vibe-generated Pester tests for new rules
+
+### Fixed
+- `Measure-AvoidSimpleFunctions` now excludes class ctors, class methods, filters, and workspace functions
+- `Measure-AvoidSimpleFunctions` no longer replaces the entire function extent -- only the first line
+- `Measure-AvoidLongTypeNames` now determines the `using namespace` correction extent based on the larger of existing `#requires` and `using` entries
+
+## [0.4.3] - 2025-10-08
+
+### Added
+- Added `Measure-AvoidSimpleFunctions` rule refactored from VSCodeProfile
+- Added `findEditorSimpleFunctions` private function refactored from VSCodeProfile
+- Added `Measure-CheckParamBlockParen` rule to insert space between `param` keywords and parentheses
 
 ## [0.3.3] - 2025-10-07